Showcases how to append something from the server using turbo-streams.
It was a bit complicated to make Thymeleaf work. I ended up registering a ThymeleafViewResolver that handles only templates with `stream` on their names. Other templates use the default view resolver. This was necessary to make turbo-frame and turbo-stream work together.

This creates the need for the developer to remember to name a template with `stream` in part of the name if one wants to use it as a turbo-stream.

TODO: The ideal solution would be to check if Turbo sent the `turbo-stream` on the Accept header and change the resolver in runtime accordingly.